feat: add sandbox configuration support and update CLI to v2.0.57 (v0.1.10)
- Add SandboxSettings, SandboxNetworkConfig, SandboxIgnoreViolations types
- Add WithSandbox option for configuring bash command isolation
- Merge sandbox settings into CLI settings when provided
- Add stream close timeout handling for SDK MCP servers
- Change HookMatcher.Timeout type from *int to *float64 for Python SDK consistency
- Update bundled CLI version to 2.0.57

+++ internal/shared/control.go
@@ -91,7 +91,7 @@ type InitializeRequest struct {
 type HookMatcherConfig struct {
 	Matcher         string   `json:"matcher"`
 	HookCallbackIDs []string `json:"hookCallbackIds"`
-	Timeout         *int     `json:"timeout,omitempty"`
+	Timeout         *float64 `json:"timeout,omitempty"`
 }
 
 // CanUseToolRequest represents a tool permission request from CLI.
